backend/requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
google-cloud-firestore==2.13.1
google-cloud-scheduler==2.14.0
openai==1.35.1
httpx>=0.25.0,<0.28.0
python-dotenv==1.0.0
flask==2.3.3
flask-cors==4.0.0
firebase-admin==6.2.0
